BGDIINF_SB-2131: Fixed missing results
When comparing the following search results

/rest/services/ech/SearchServer?sr=2056&searchText=Be&lang=de&type=locations

to the results from the actual productive service, the new one had some
missing results and the other results had a weight of -1 in comparaison
to the prod results.

This was due to some porting mistakes (typos) and a missing part of the
search functionality.