Removed a bad style broad exception
The broad exception has been replaced by the eventuel possible
exception. In the try ... except block, the method `box()` and `Point()`
could trigger a TypeError exception if they did receive a wrong type
arguments but this should not happen, that's why this exception is not
catched. If this would happen we would find this out more quickly as the
request would failed with a 500.
In opposite if the results is an empty string (which apparently can
happen and is ok), the method `parse_box2d()` would raise a ValueError.
In this case we catch the exception, log an error an bail out with True.