BGDIINF_SB-2126: Fixed callback feature and clean up
The callback functionality did not contained a valid json content; used
single quote instead of double quotes and special caracters where not
escaped correctly, this was due to the fact that the python dictionary
was printed instead of a json dump. Also removed superflous `)` the
callback answer, this was making it an invalid javascript syntax.

Clean up legacy Paramid framework variable `matchdict` that came from
the legacy mf-chsdi3.