BGDIINF_SB-2299: Improve performance by factor 6x
The service was massively slower than the original from mf-chsdi3.

I found out that the function transform_shape was very slow and that the cause was
the pyproj library version that we used. Upgrading pyproj from 3.0.0 to 3.3.0 improved
the performance by a factor 6x !
I also update other libraries to the latest. Those ones however don't have an impact on performance.

The transform_shape() still has some performance issue though, using a simple request showed that the
method still needs 0.045s and because depending on the result this can be called several times it can
lead to poor performance (e.g. in the test that used it was called 30 times => ~1.5s, while doing
the same test on the service running mf-chsdi3 took 0.4s)