BGDIINF_SB-2299: Fixed simple cache error
The simple cache of flask cannot return complex object nor be used with method that
takes complex object as argument.

Therefore now the `get_transformer()` function use a global variable to cache the
result. Here we don't need to have a timeout as the transformer should never change.
Also as we have a finite combination of SRID from-to we are safe by saving each combination
into a global variable.