BGDIINF_SB-2322: Improved exception handling
During performance tests we had a 504 Gateway timeout, triggered from sphinx
server. Based on this error the logging was quite poor and we were lacking of
the backtrace. So now in case of a 5XX response we use the `logger.exception`
which sets the backtrace in the log message. Backtrace often helps to
understand the reason and cause of failure.