BGDIINF_SB-2322: Made the sphinxapi timeout configurable
While doing performance tests, we had a one time failure due to this timeout
of 10 seconds. This failure could unfortunately not be explained. Neither the
reason of such high timeout could be (originally it was 3.5 seconds but has
been changed to 10s in 2014 without explanation).

So I made it configurable but let it to 3s as default. 3 seconds is a much more
realistic value, as a user usually don't want to wait 10 seconds on a search
request ! If we stumble again on the root cause of having 10 seconds as a timeout
we can still easily change it back to 10 by changing the service configuration
without re-creating a new service release.