BGDIINF_SB-2322: Sets the wsgi default workers to the double of CPU
By default use the number of CPU to set the wsgi workers, however this value
can be set to a fix value by configuration settings.

Usually the number of workers depends on the deployment, for example on k8s
scalling is done horizontally by k8s, while on vhost scaling is done vertically.