BGDIINF_SB-2322: Improved error handling
During a service upgrade with only one Node, we had some failures (NoneType
exception) due to unavailable sphinx server which was totally up.

So now the error handling has been improved also returning an appropriate 503
when the sphinx server is not ready. Logging of error has also been improved
to help to debug in such case.