BGDIINF_SB-3058: Fixed transform geometry caching collision
The `cache.memoize()` method uses `repr` for object argument which can creates
cache key collision because the first parameter is an Polygon object.
A Polygon object is represented as follow with `repr`:

`<shapely.geometry.polygon.Polygon object at 0x7f23f6848e80>`.
The hex number being the memory address of the object.

It can happen that two different polygon objects reuse the same memory address
which results to the same cache key for two different polygons ! This is what
happen when searching for `blauseeli` but was a pure hazard and was not easy
to reproduce locally.

The fix is to use `cached()` with a custom cache key maker which uses `str`
instead of `repr` for the polygon object. This represent the polygon as
follow:

`POLYGON ((1 1, 1 1, 1 1, 1 1, 1 1))`