PB-299: when doing queries on the swisssearch distributed index which consists
of the following parts:
    type = distributed
    local = zipcode         -> rank 1
    local = district        -> rank 3
    local = kantone         -> rank 4
    local = gg25            -> rank 2
    local = swissnames3d    -> rank 5
    local = haltestellen    -> rank 6
    local = parcel          -> rank 10
    local = address         -> rank 7

not all parts of the index are returned in the results when the request
is sent without origins parameter which is the default.

The coverage of all the parts of the distributed index in the result can
be forced by setting the filter to all the ranks that are available.

NOTE: it is unknown why the problem has been occurring since 18/02/2024.