PB-1167: create a custom query string for the fuzzy search
when searching in the metaphone index a simple searchText is sufficient,
there is no need for all the wildcard variants.

digit/text combinations will be expanded to (digit|digit/text) for a
better support of typos.

p.e.
['seestrasse', '78b', 'uster'] will be converted into
@detail seestrasse (78|78b) uster

the quorum operator will be used for the support of fuzzy matching. as a
starting value the quorum operor is matching documents with 70% overlap
( 70% of the search text keywords are present in the document )